requirements.txt
asgiref==3.8.1
astroid==3.3.4
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
dill==0.3.9
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.1.0
flake8==7.1.1
inflection==0.5.1
ipython==8.27.0
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.7.0
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.1
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomlkit==0.13.2
traitlets==5.14.3
tzdata==2024.1
uritemplate==4.1.1
wcwidth==0.2.13
